server/requirements.txt
grpcio
grpcio-tools
protobuf
flask
flask-socketio
flask-cors

